Freshen the GUI
Everything is smaller, shorter, tidier, more compact and mostly made of rem. Things should shrink to a minimum, without text wrapping. An effort has been made so that it could be usable on mobile, but don't strain your eyes.
There are slight accents of purple and green here and there as a sort of theme. I've also moved and renamed things, and I'm not sure if more can be done. I'm particularly not familiar with good commenting and naming practices.
I've also taken the liberty of changing the two logos to their no-text versions, because I think the text was redundant.

I tried my best to improve it as it is, I don't mind if it's rejected in full, and I'm open to suggestions to improve/fix it.
It definitely needs review to double check I didn't break anything. If needed, I will provide screenshots/videos for a faster process.